Check for active servers
It is a one liner that gets the JSON of Cloud Data from the Status Scratch Project and then maps the array there to a true/false array based on whether the variable has been updated correctly and then it is checks if there are any "true" values and then it reports accordingly. All in one line because it takes up less space.